// src/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */
import {QueryFormData} from '@superset-ui/core';
import EChartsReact, {EChartsOption} from 'echarts-for-react';

export interface EchartsChinaMapStylesProps {
  height: number;
  width: number;
}

export interface EchartsChinaMapState {
  echartRef?: EChartsReact;
  chartOption: EChartsOption;
  isLoading: boolean;
  currentPrefix: string;
  currentAdcode: string;
  metricName: string;
}

export interface EchartsChinaMapDataRecord {
  adcode: string;
  value: number;
}

export interface EchartsChinaMapDataRecordPrefixed extends EchartsChinaMapDataRecord {
  name?: string;
  prefix: string;
}

interface EchartsChinaMapCustomizeProps {
}

export type EchartsChinaMapQueryFormData = QueryFormData &
    EchartsChinaMapStylesProps &
    EchartsChinaMapCustomizeProps;


export type EchartsChinaMapProps = EchartsChinaMapStylesProps &
    EchartsChinaMapCustomizeProps & {
  data: EchartsChinaMapDataRecord[];
  metricName: string;
};